chore: download progress finished should reload model list
